Move cv::Mat parser into wave_vision
