+++ wave_vision/tests/data/config.yaml
@@ -1,0 +1,6 @@
+test_matrix:
+  rows: 3
+  cols: 4
+  data: [1.1, 2.2, 3.3, 4.4,
+         5.5, 6.6, 7.7, 8.8,
+         9.9, 11.0, 12.1, 13.2]
